+++ .gitignore
@@ -33,7 +33,7 @@ x86/
 [Aa][Rr][Mm]/
 [Aa][Rr][Mm]64/
 bld/
-album-api/[Bb]in
+albumapi/[Bb]in
 # [Bb]in/
 [Oo]bj/
 [Ll]og/
