+++ .vscode/launch.json
@@ -14,7 +14,7 @@
       "name": "Album Viewer",
       "type": "node",
       "request": "launch",
-      "cwd": "${workspaceFolder}/album-viewer",
+      "cwd": "${workspaceFolder}/albumviewer",
       "runtimeExecutable": "npm",
       "skipFiles": ["<node_internals>/**"],
       "runtimeArgs": ["run-script", "start"],
@@ -24,7 +24,7 @@
         "DAPR_HTTP_PORT": "3180",
         "DAPR_GRPC_PORT": "3101",
         "DAPR_METRICS_PORT": "9091",
-        "ALBUM_API": "album-api",
+        "albumapi": "albumapi",
         "PORT": "3000"
       },
       "console": "integratedTerminal",
@@ -40,9 +40,9 @@
       "request": "launch",
       "preLaunchTask": "Dapr up: Album API",
       "postDebugTask": "Dapr down: Album API",
-      "program": "${workspaceFolder}/album-api/bin/Debug/net6.0/album-api.dll",
+      "program": "${workspaceFolder}/albumapi/bin/Debug/net6.0/albumapi.dll",
       "args": [],
-      "cwd": "${workspaceFolder}/album-api",
+      "cwd": "${workspaceFolder}/albumapi",
       "stopAtEntry": false,
       "env": {
         "ASPNETCORE_ENVIRONMENT": "Development",
