+++ albumviewer/bin/www
@@ -5,7 +5,7 @@
  */
 
 var app = require("../app");
-var debug = require("debug")("album-viewer:server");
+var debug = require("debug")("albumviewer:server");
 var http = require("http");
 
 /**
