+++ azure.yaml
@@ -1,9 +1,15 @@
 # yaml-language-server: $schema=https://raw.githubusercontent.com/Azure/azure-dev/main/schemas/v1.0/azure.yaml.json
 
-name: containerapps-dapralbums
-infra:
-    provider: ""
-    path: ""
-    module: ""
-pipeline:
-    provider: ""
+name: DaprAlbums
+services:
+  albumapi:
+    project: albumapi
+    dist: build
+    language: csharp
+    host: containerapp
+    module: app/albumapi
+  albumviewer:
+    project: albumviewer
+    language: js
+    host: containerapp
+    module: app/albumviewer
