+++ deploy/local-dev/build.sh
@@ -1,10 +1,10 @@
 # restore and build .NET project
-cd album-api
+cd albumapi
 dotnet restore
 dotnet build
 cd ..
 
 # npm install Node.js app
-cd album-viewer
+cd albumviewer
 npm install
 cd ..
