+++ infra/core/host/container-apps-environment.bicep
@@ -2,7 +2,10 @@ param environmentName string
 param location string = resourceGroup().location
 
 param containerAppsEnvironmentName string = ''
-param logAnalyticsWorkspaceName string
+param logAnalyticsWorkspaceName string = ''
+param daprAIInstrumentationKey string = ''
+param vnetInfrastructureSubnetId string = ''
+param vnetIsInternal bool = false
 
 var abbrs = loadJsonContent('../../abbreviations.json')
 var resourceToken = toLower(uniqueString(subscription().id, environmentName, location))
@@ -13,6 +16,11 @@ resource containerAppsEnvironment 'Microsoft.App/managedEnvironments@2022-03-01'
   location: location
   tags: tags
   properties: {
+    daprAIInstrumentationKey: !empty(daprAIInstrumentationKey) ? daprAIInstrumentationKey : null
+    vnetConfiguration: empty(vnetInfrastructureSubnetId) ? null : {
+      internal: vnetIsInternal
+      infrastructureSubnetId: vnetInfrastructureSubnetId
+    }
     appLogsConfiguration: {
       destination: 'log-analytics'
       logAnalyticsConfiguration: {
