+++ infra/core/host/container-apps.bicep
@@ -6,6 +6,9 @@ param containerAppsGroupName string = 'app'
 param containerRegistryName string = ''
 param containerRegistrySku object = { name: 'Standard' }
 param logAnalyticsWorkspaceName string = ''
+param daprAIInstrumentationKey string = ''
+param vnetInfrastructureSubnetId string = ''
+param vnetIsInternal bool = false
 
 module containerAppsEnvironment 'container-apps-environment.bicep' = {
   name: '${containerAppsGroupName}-container-apps-environment'
@@ -14,6 +17,9 @@ module containerAppsEnvironment 'container-apps-environment.bicep' = {
     location: location
     containerAppsEnvironmentName: containerAppsEnvironmentName
     logAnalyticsWorkspaceName: logAnalyticsWorkspaceName
+    daprAIInstrumentationKey: daprAIInstrumentationKey
+    vnetInfrastructureSubnetId: vnetInfrastructureSubnetId
+    vnetIsInternal: vnetIsInternal
   }
 }
