+++ infra/core/host/dapr-statestore.bicep
@@ -30,6 +30,6 @@ resource daprComponent 'Microsoft.App/managedEnvironments/daprComponents@2022-06
       }
     ]
     secretStoreComponent: secretStoreName
-    scopes: ['album-api']
+    scopes: ['albumapi']
   }
 }
