+++ infra/core/security/keyvault-secret.bicep
@@ -1,0 +1,23 @@
+param environmentName string
+param location string = resourceGroup().location
+
+param secretName string
+@secure()
+param secretValue string
+param keyVaultName string = ''
+
+var abbrs = loadJsonContent('../../abbreviations.json')
+var resourceToken = toLower(uniqueString(subscription().id, environmentName, location))
+
+resource keyVault 'Microsoft.KeyVault/vaults@2022-07-01' existing = {
+  name: !empty(keyVaultName) ? keyVaultName : '${abbrs.keyVaultVaults}${resourceToken}'
+}
+
+resource secret 'Microsoft.KeyVault/vaults/secrets@2021-11-01-preview' = {
+  parent: keyVault
+  name: secretName
+  properties: {
+    value: secretValue
+  }
+}
+
